fix(homepage): hide the play store banner on ios and share the CTA style
PR review feedback:

- AppPromoBanner is now gated on `isIOS` (react-device-detect, already used in
  SideMenu). useCanPromoteApp only tells us the visitor is not already in the
  installed app, so iOS Safari users were shown a Google Play badge for a store
  that does not exist on their platform.
- The pill shape + hover lift shared by the three home page CTAs moves to a
  single AttractiveButton, so they cannot drift apart.
- The Play Store badge is lazy-loaded and the comment now mentions the external
  request to Google's CDN.
- Documented the iOS gate living inside PWAPrompt and the untranslated
  copySubtitle.
