+++ packages/web-app/src/pages/homepage/IosInstallPrompt.jsx
@@ -5,6 +5,11 @@ import PWAPrompt from 'react-ios-pwa-prompt';
 // beforeinstallprompt event, so the only way to install the PWA is via
 // Share -> Add to Home Screen. The library self-gates on iOS + non-standalone
 // and throttles reappearance via localStorage.
+//
+// Note there is deliberately no platform check here: PWAPrompt is rendered
+// unconditionally and decides internally whether to show anything. It still
+// mounts and reads localStorage on every platform, which is cheap enough not
+// to be worth a second gate — the iOS check is inside the library, not missing.
 const IosInstallPrompt = () => {
   const { formatMessage } = useIntl();
 
@@ -15,6 +20,7 @@ const IosInstallPrompt = () => {
       copyDescription={formatMessage({
         id: 'Install Grottocenter to browse caves offline and in fullscreen, right in the field.'
       })}
+      // Not translated on purpose: this is the site's domain, not UI copy.
       copySubtitle="grottocenter.org"
       copyShareStep={formatMessage({
         id: "Tap the 'Share' button in the toolbar below"
